+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [2.3.0] - 2025-11-20
+
+This release improves the CLI tool to make it more useful for server maintenance and scripting.
+
+### Added
+
+- Added commands to the CLI tool:
+  - `install`: Print parameters for installing outdated packages for usage with `pip install`.
+  - `outdated`: Show outdated distribution packages
+  - `version`: Show version of package_manager app
+
 ## [2.2.2] - 2025-05-26
 
 ### Fixed
@@ -217,7 +228,7 @@ To address this, such potentially unwanted updates can now be automatically excl
 
 ### Fixed
 
-- Fetching info from PyPI for a package breaks when requires_python contains an invalid specifer
+- Fetching info from PyPI for a package breaks when requires_python contains an invalid specifier
 
 ## [1.8.0] - 2023-01-03